django_project/requirements.txt
django
Pillow
dnspython
enum-compat
eventlet
greenlet
monotonic
python-engineio
python-socketio
pytz
six
psycopg2
djangorestframework
pyJWT
djangorestframework-simplejwt
django-phonenumber-field[phonenumbers]
django-countries
django-cities
django-cors-headers
scikit-fuzzy
numpy
pandas
matplotlib
scikit-learn
pickle-mixin
vonage